WIP: clean up tests
