+++ shed/tests/test_streams.py
@@ -14,44 +14,16 @@
 ##############################################################################
 # TODO: this should be broken up into a module per node, that way we
 # can test individual nodes without having to test everything
-from numpy.testing import assert_allclose, assert_equal, assert_raises
+from itertools import zip_longest
+
+import pytest
+from numpy.testing import assert_allclose, assert_equal
 from streamz.core import Stream
 
 import shed.event_streams as es
+from shed.tests.utils import SinkAssertion
 from ..event_streams import dstar, star
-import pytest
-from bluesky.callbacks.core import CallbackBase
-from itertools import zip_longest
 from ..utils import to_event_model
-
-
-class SinkAssertion(CallbackBase):
-    def __init__(self, fail=True, expected_docs=None):
-        self.fail = fail
-        self.docs = []
-        if expected_docs is None:
-            if fail:
-                self.expected_docs = {'start', 'descriptor', 'stop'}
-            else:
-                self.expected_docs = {'start', 'descriptor', 'event', 'stop'}
-        else:
-            self.expected_docs = expected_docs
-
-    def __call__(self, name, doc):
-        """Dispatch to methods expecting particular doc types."""
-        self.docs.append(name)
-        return getattr(self, name)(doc)
-
-    def stop(self, doc):
-        if self.fail:
-            assert doc['exit_status'] == 'failure'
-            assert doc.get('reason')
-        else:
-            assert doc['exit_status']
-            if not doc.get('reason', None):
-                print(doc.get('reason', None))
-            assert not doc.get('reason', None)
-        assert self.expected_docs == set(self.docs)
 
 
 def test_clear():
@@ -62,476 +34,6 @@ def test_clear():
     assert s.md == {'hello': 'globe'}
     s._clear()
     assert s.md == {'hello': 'world'}
-
-
-def test_map(exp_db, start_uid1):
-    source = Stream()
-
-    def add5(img):
-        return img + 5
-
-    ii = {'img': 'pe1_image'}
-    oi = [('image', {'dtype': 'array', 'source': 'testing'})]
-    dp = es.map(add5,
-                source,
-                input_info=ii,
-                output_info=oi, stream_name='test')
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    prov = dict(stream_class='map',
-                function=dict(function_module=add5.__module__,
-                              function_name=add5.__name__),
-                stream_class_module=es.map.__module__,
-                input_info=ii, output_info=oi)
-    assert_docs = set()
-    for l, s in zip(L, exp_db.restream(ih1, fill=True)):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            assert_allclose(l[1]['data']['image'],
-                            s[1]['data']['pe1_image'] + 5)
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-            assert l[1]['provenance'] == prov
-        assert l[1] != s[1]
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_map_nested_input_info(exp_db, start_uid1):
-    source = Stream()
-
-    def add5(img):
-        return img + 5
-
-    ii = {'img': (('data', 'pe1_image'), 0)}
-    oi = [('image', {'dtype': 'array', 'source': 'testing'})]
-    dp = es.map(add5,
-                source,
-                input_info=ii,
-                output_info=oi, stream_name='test')
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    prov = dict(stream_class='map',
-                function=dict(function_module=add5.__module__,
-                              function_name=add5.__name__),
-                stream_class_module=es.map.__module__,
-                input_info=ii, output_info=oi)
-    assert_docs = set()
-    for l, s in zip(L, exp_db.restream(ih1, fill=True)):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            assert_allclose(l[1]['data']['image'],
-                            s[1]['data']['pe1_image'] + 5)
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-            assert l[1]['provenance'] == prov
-        assert l[1] != s[1]
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_map_no_input_info(exp_db, start_uid1):
-    source = Stream()
-
-    dp = es.map(lambda **x: x, source)
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-    dp.sink(print)
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=False)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    for l, s in zip(L, exp_db.restream(ih1, fill=False)):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            assert l[1]['data'] == s[1]['data']
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-        assert l[1] != s[1]
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_map_no_input_info_full_event(exp_db, start_uid1):
-    source = Stream()
-
-    dp = es.map(lambda **x: x, source, full_event=True)
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-    dp.sink(print)
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=False)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    for l, s in zip(L, exp_db.restream(ih1, fill=False)):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            assert l[1]['data'] == s[1]
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-        assert l[1] != s[1]
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_map_args(exp_db, start_uid1):
-    source = Stream()
-
-    from operator import add
-
-    ii = {0: 'pe1_image'}
-    oi = [('image', {'dtype': 'array', 'source': 'testing'})]
-    dp = es.map(add,
-                source,
-                5,
-                input_info=ii,
-                output_info=oi, md={'stream_name': 'test'})
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    prov = dict(stream_class='map',
-                function=dict(function_module=add.__module__,
-                              function_name=add.__name__),
-                stream_class_module=es.map.__module__,
-                input_info=ii, output_info=oi)
-    assert_docs = set()
-    for l, s in zip(L, exp_db.restream(ih1, fill=True)):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            assert_allclose(l[1]['data']['image'],
-                            s[1]['data']['pe1_image'] + 5)
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-            assert l[1]['provenance'] == prov
-        assert l[1] != s[1]
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_map_args_kwargs(exp_db, start_uid1):
-    source = Stream()
-
-    def add(img, adder):
-        return img + adder
-
-    ii = {0: 'pe1_image'}
-    oi = [('image', {'dtype': 'array', 'source': 'testing'})]
-    dp = es.map(add,
-                source,
-                adder=5,
-                input_info=ii,
-                output_info=oi)
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    prov = dict(stream_class='map',
-                function=dict(function_module=add.__module__,
-                              function_name=add.__name__),
-                stream_class_module=es.map.__module__,
-                input_info=ii, output_info=oi)
-    assert_docs = set()
-    for l, s in zip(L, exp_db.restream(ih1, fill=True)):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            assert_allclose(l[1]['data']['image'],
-                            s[1]['data']['pe1_image'] + 5)
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-            assert l[1]['provenance'] == prov
-        assert l[1] != s[1]
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_map_two_runs(exp_db, start_uid1):
-    source = Stream()
-
-    def add5(img):
-        return img + 5
-
-    ii = {'img': 'pe1_image'}
-    oi = [('image', {'dtype': 'array', 'source': 'testing'})]
-    dp = es.map(add5,
-                source,
-                input_info=ii,
-                output_info=oi)
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    prov = dict(stream_class='map',
-                function=dict(function_module=add5.__module__,
-                              function_name=add5.__name__),
-                stream_class_module=es.map.__module__,
-                input_info=ii, output_info=oi)
-    assert_docs = set()
-    for l, s in zip(L, exp_db.restream(ih1, fill=True)):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            assert_allclose(l[1]['data']['image'],
-                            s[1]['data']['pe1_image'] + 5)
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-            assert l[1]['provenance'] == prov
-        assert l[1] != s[1]
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-    L_original = L.copy()
-    del L[:]
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    for l, s, ll in zip(L, exp_db.restream(ih1, fill=True), L_original):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            assert_allclose(l[1]['data']['image'],
-                            ll[1]['data']['image'])
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-            assert l[1]['provenance'] == ll[1]['provenance']
-        assert l[1] != s[1]
-        assert_raises(AssertionError, assert_equal, l[1], ll[1])
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_map_full_event(exp_db, start_uid1):
-    source = Stream()
-
-    def add5(i):
-        return i + 5
-
-    ii = {'i': 'seq_num'}
-    oi = [('i', {'dtype': 'array', 'source': 'testing'})]
-    dp = es.map(add5,
-                source,
-                input_info=ii,
-                output_info=oi,
-                full_event=True)
-    dp.sink(star(SinkAssertion(False)))
-    L = dp.sink_to_list()
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    prov = dict(stream_class='map',
-                function=dict(function_module=add5.__module__,
-                              function_name=add5.__name__),
-                stream_class_module=es.map.__module__,
-                input_info=ii, output_info=oi,
-                full_event=True)
-
-    assert_docs = set()
-    for l, s in zip(L, exp_db.restream(ih1, fill=True)):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            print(l[1])
-            print(s[1]['seq_num'])
-            assert_allclose(l[1]['data']['i'],
-                            s[1]['seq_num'] + 5)
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-            assert l[1]['provenance'] == prov
-        assert l[1] != s[1]
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_map_stream_input(exp_db, start_uid1):
-    source = Stream()
-
-    def add5(img):
-        return img + 5
-
-    ii = {'img': ('pe1_image', source)}
-    oi = [('img', {'dtype': 'array', 'source': 'testing'})]
-    dp = es.map((add5),
-                source,
-                input_info=ii,
-                output_info=oi)
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    prov = dict(stream_class='map',
-                function=dict(function_module=add5.__module__,
-                              function_name=add5.__name__),
-                stream_class_module=es.map.__module__,
-                input_info=ii, output_info=oi)
-
-    assert_docs = set()
-    for l, s in zip(L, exp_db.restream(ih1, fill=True)):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            assert_allclose(l[1]['data']['img'], s[1]['data']['pe1_image'] + 5)
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-            assert l[1]['provenance'] == prov
-        assert l[1] != s[1]
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_double_map(exp_db, start_uid1):
-    source = Stream()
-    source2 = Stream()
-
-    def add_imgs(img1, img2):
-        return img1 + img2
-
-    dp = es.map((add_imgs), es.zip(source, source2),
-                input_info={'img1': ('pe1_image', 0),
-                            'img2': ('pe1_image', 1)},
-                output_info=[
-                    ('img',
-                     {'dtype': 'array',
-                      'source': 'testing'})])
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-        source2.emit(a)
-
-    assert_docs = set()
-    for l, s in zip(L, exp_db.restream(ih1, fill=True)):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            assert_allclose(l[1]['data']['img'],
-                            add_imgs(s[1]['data']['pe1_image'],
-                                     s[1]['data']['pe1_image']))
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_double_internal_map(exp_db, start_uid1):
-    source = Stream()
-
-    def div(img1, ct):
-        return img1 / ct
-
-    dp = es.map(div, source,
-                input_info={'img1': 'pe1_image', 'ct': 'I0'},
-                output_info=[
-                    ('img',
-                     {'dtype': 'array',
-                      'source': 'testing'})])
-
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion(False)))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    for l, s in zip(L, exp_db.restream(ih1, fill=True)):
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            assert_allclose(l[1]['data']['img'], div(s[1]['data']['pe1_image'],
-                                                     s[1]['data']['I0']))
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-@pytest.mark.xfail(raises=TypeError)
-def test_map_fail(exp_db, start_uid1):
-    source = Stream()
-
-    def add5(img):
-        return img + 5
-
-    ii = {'i': 'pe1_image'}
-    oi = [('image', {'dtype': 'array', 'source': 'testing'})]
-    dp = es.map((add5),
-                source,
-                input_info=ii,
-                output_info=oi)
-    dp.sink(star(SinkAssertion()))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-
-def test_map_fail_dont_except(exp_db, start_uid1):
-    source = Stream()
-
-    def add5(img):
-        return img + 5
-
-    ii = {'i': 'pe1_image'}
-    oi = [('image', {'dtype': 'array', 'source': 'testing'})]
-    dp = es.map(add5,
-                source,
-                input_info=ii,
-                output_info=oi, raise_upon_error=False)
-    L = dp.sink_to_list()
-    dp.sink(star(SinkAssertion()))
-
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    for l, s in zip(L, exp_db.restream(ih1, fill=True)):
-        assert_docs.add(l[0])
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'failure'
-        assert l[1] != s[1]
-    for n in ['start', 'descriptor', 'stop']:
-        assert n in assert_docs
 
 
 def test_filter(exp_db, start_uid1):
