+++ shed/tests/utils.py
@@ -18,6 +18,7 @@ from uuid import uuid4
 import numpy as np
 from bluesky.examples import ReaderWithRegistry, Reader
 from bluesky.plans import count
+from bluesky.callbacks.core import CallbackBase
 
 
 def insert_imgs(RE, reg, n, shape, save_dir=tempfile.mkdtemp(), **kwargs):
@@ -61,3 +62,32 @@ def insert_imgs(RE, reg, n, shape, save_dir=tempfile.mkdtemp(), **kwargs):
     RE(count([light_det, light_i0], num=n), **light_md)
 
     return save_dir
+
+
+class SinkAssertion(CallbackBase):
+    def __init__(self, fail=True, expected_docs=None):
+        self.fail = fail
+        self.docs = []
+        if expected_docs is None:
+            if fail:
+                self.expected_docs = {'start', 'descriptor', 'stop'}
+            else:
+                self.expected_docs = {'start', 'descriptor', 'event', 'stop'}
+        else:
+            self.expected_docs = expected_docs
+
+    def __call__(self, name, doc):
+        """Dispatch to methods expecting particular doc types."""
+        self.docs.append(name)
+        return getattr(self, name)(doc)
+
+    def stop(self, doc):
+        if self.fail:
+            assert doc['exit_status'] == 'failure'
+            assert doc.get('reason')
+        else:
+            assert doc['exit_status']
+            if not doc.get('reason', None):
+                print(doc.get('reason', None))
+            assert not doc.get('reason', None)
+        assert self.expected_docs == set(self.docs)
